Other Files/api Manager/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0
openpyxl>=3.1.0
xlrd>=2.0.0
dropbox>=11.36.0
pydantic>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3
pint>=0.22  # For unit conversions